+++ R/preprocess-landings.R
@@ -28,7 +28,7 @@
 #' Progress through the function is tracked using the package *logger*.
 #'
 #' @param log_threshold
-#' @inheritParams ingest_updated_landings
+#' @inheritParams ingest_landings
 #' @keywords workflow
 #' @return no outputs. This function is used for it's side effects
 #' @export
@@ -333,7 +333,7 @@ preprocess_landings_step_2 <- function(log_threshold = logger::DEBUG) {
 #' Progress through the function is tracked using the package *logger*.
 #'
 #' @param log_threshold
-#' @inheritParams ingest_legacy_landings
+#' @inheritParams ingest_landings
 #' @keywords workflow
 #' @return no outputs. This funcrion is used for it's side effects
 #' @export