+++ R/retrieve-survey-data.R
@@ -1,263 +1,131 @@
-#' Select server from kobo
+#' Retrieve Data from Kobotoolbox API
 #'
-#' Specifies the Host URL of the API to use
-#'
-#' @param api Either "kobo", "kobohr", "ona", or a custom (full) URL. API URLs
-#'   are made available for KoBo Toolbox ("kobo",
-#'   \url{https://kc.kobotoolbox.org/api/v1/}), KoBo Humanitarian Response
-#'   ("kobohr", \url{https://kc.humanitarianresponse.info/api/v1/}), Ona ("ona",
-#'   \url{https://api.ona.io/api/v1/}) and Unhcr ("unhcr",
-#'   \url{https://kobocat.unhcr.org/api/v1/}) . For your own installation, or
-#'   other installations using the same API but accessed at a different URL,
-#'   enter the full URL.
-#' @param version Wether using the API "v1" or "v2"
-#'
-#' @author Ananda Mahto
-#'
-#' @return An URL
-#' @export
-#'
-#' @examples
-#'
-#' get_host_url("unhcr")
-#' get_host_url("https://kobocat.unhcr.org/api/v1/")
-get_host_url <- function(api, version = "v1") {
-  if (api %in% c("kobo", "kobohr", "ona", "unhcr")) {
-    if (version == "v1") {
-      switch(api,
-        kobo = "https://kc.kobotoolbox.org/api/v1",
-        kobohr = "https://kc.humanitarianresponse.info/api/v1",
-        ona = "https://api.ona.io/api/v1",
-        unhcr = "https://kobocat.unhcr.org/api/v1"
-      )
-    } else if (version == "v2") {
-      switch(api,
-        kobo = "https://kc.kobotoolbox.org/api/v2",
-        kobohr = "https://kobo.humanitarianresponse.info/api/v2"
-      )
-    }
-  } else {
-    api
-  }
-}
-
-#' Download kobo data in as json using the v1 API
-#'
-#' @param path string with path to file where API request should be saved
-#' @param id survey id. Usually a 6 digit number. See [this support
-#'   page](https://support.kobotoolbox.org/pulling_data_into_powerbi.html) for
-#'   an example on how this can be obtained
-#' @param token access token for the account e.g. "Token XXXXXXX"
-#' @param api
-#' @param overwrite Will only overwrite existing path if TRUE.
-
-#'
-#' @inheritParams get_host_url
-#'
-#' @return The file path
-#' @export
-#'
-#' @examples
-#' \dontrun{
-#' retrieve_survey_data("test.json", id = 753491, token = "XXX")
-#' file.remove("test.json")
-#' }
-retrieve_survey_data <- function(path, id = NULL, token = NULL,
-                                 api = "kobohr",
-                                 overwrite = TRUE) {
-  api_limit <- 30000
-
-  request_url <- paste(get_host_url(api, "v1"),
-    "data", id,
-    sep = "/"
-  )
-
-  message(request_url)
-
-  # Function to get a page of survey
-  retrieve_survey_page <- function(start = 0) {
-    message("Retrieving survey. Starting at record ", start)
-
-    resp <- httr::RETRY(
-      verb = "GET",
-      url = request_url,
-      config = httr::add_headers(Authorization = token),
-      query = list(
-        start = start,
-        limit = api_limit
-      ),
-      times = 12,
-      pause_cap = 180
-    )
-
-    if (!(resp$status_code %in% 200:299)) {
-      stop("Unsuccessful response from server")
-    }
-
-    content <- httr::content(resp)
-    if ("results" %in% names(content)) content <- content$results
-    content
-  }
-
-  results <- list()
-  get_next <- TRUE
-  start <- 0
-
-  while (get_next) {
-    this_page <- retrieve_survey_page(start)
-    results <- c(results, this_page)
-    if (length(this_page) < api_limit) {
-      get_next <- FALSE
-    } else {
-      start <- start + api_limit
-    }
-  }
-
-  # Check that submissions are unique in case there is overlap in the pagination
-  if (dplyr::n_distinct(purrr::map_dbl(results, ~ .$`_id`)) != length(results)) {
-    stop("Number of submission ids not the same as number of records")
-  }
-
-  jsonlite::write_json(results, path, auto_unbox = TRUE)
-  path
-}
-
-#' Download survey metadata
-#'
-#' @param id
-#' @param token
-#' @param api
-#' @inheritParams retrieve_survey_data
-#'
-#' @return A list with survey metadata
-#' @export
-#'
-#' @examples
-#' \dontrun{
-#' # It's only possible to donwload survey metadata with the account Token
-#' retrieve_survey_metadata(753491, token = "123XXXXXX")
-#' }
-#'
-retrieve_survey_metadata <- function(id = NULL, token = NULL, api = "kobohr") {
-  # If the token is not provided is not possible to get the
-  no_token <- length(token) == 0 | is.null(token)
-  if (no_token) {
-    stop("Token is required")
-  }
-
-  survey_basic_metadata <- retrieve_basic_survey_metadata(id, token, api)
-
-  # the version 2 of the api returns much richer information about the surveys
-  httr::GET(
-    url = paste(get_host_url(api, version = "v2"), "assets",
-      survey_basic_metadata$id_string,
-      sep = "/"
-    ),
-    config = httr::add_headers(Authorization = token)
-  ) %>%
-    httr::content() %>%
-    c(survey_basic_metadata)
-}
-
-# Help function to retrieve basic metadata from the v1 api - Same params as in
-# retrieve_survey_metadata
-retrieve_basic_survey_metadata <- function(id, token, api) {
-  assets_v1_raw <- httr::GET(
-    url = paste(get_host_url(api, version = "v1"), "data", sep = "/"),
-    config = httr::add_headers(Authorization = token)
-  )
-
-  purrr::keep(httr::content(assets_v1_raw), ~ .$id == id)[[1]]
-}
-
-
-#' Download survey
-#'
-#' Download multiple survey artefacts (data and metadata) at once. Progress
-#' through the function is tracked using the package *logger*.
+#' This function retrieves survey data from Kobotoolbox API for a specific asset.
+#' It supports pagination and handles both JSON and XML formats.
 #'
 #' @param prefix name to be used as the prefix of the file names to be
 #'   downloaded. Can be a path.
-#' @param api
-#' @param id
-#' @param token
-#' @param format Either "csv" or "json"
-#' @param metadata whether to download metadata as well as data
-#' @param append_version whether to append versioning information to the
-#'   filename using [add_version].
-#' @inheritParams retrieve_survey_data
+#' @param assetid The asset ID of the Kobotoolbox form.
+#' @param url The URL of Kobotoolbox (default is "eu.kobotoolbox.org").
+#' @param uname Username for Kobotoolbox account.
+#' @param pwd Password for Kobotoolbox account.
+#' @param encoding Encoding to be used for data retrieval (default is "UTF-8").
+#' @param format Format of the data to retrieve, either "json" or "xml" (default is "json").
 #'
-#' @return a character vector with paths of the downloaded files
+#' @return A list containing all retrieved survey results.
+#' @keywords ingestion
+#' @details
+#' The function uses pagination to retrieve large datasets, with a limit of 30,000 records per request.
+#' It continues to fetch data until all records are retrieved or an error occurs.
+#'
 #' @export
 #'
 #' @examples
 #' \dontrun{
-#' # It's only possible to donwload survey metadata with the account Token
-#' retrieve_survey(
-#'   prefix = "my-survey", api = "kobohr", id = 753491,
-#'   token = "123XXXXXX"
-#' )
-#' # To download in a different path
-#' dir.create("my-data-dir")
-#' retrieve_survey(
-#'   prefix = "my-data-dir/my-survey", api = "kobohr", id = 753491,
-#'   token = "123XXXXXX"
+#' kobo_data <- get_kobo_data(
+#'   preifx = "my-survey",
+#'   assetid = "your_asset_id",
+#'   uname = "your_username",
+#'   pwd = "your_password"
 #' )
 #' }
-retrieve_survey <- function(prefix, api, id, token, format = c("csv", "json"),
-                            metadata = TRUE, append_version = TRUE) {
-  metadata_filename <- paste(prefix, "metadata", sep = "_")
+get_kobo_data <- function(prefix = NULL, assetid = NULL, url = "eu.kobotoolbox.org", uname = NULL, pwd = NULL, encoding = "UTF-8", format = "json") {
+  if (!is.character(url)) stop("URL entered is not a string")
+  if (!is.character(uname)) stop("uname (username) entered is not a string")
+  if (!is.character(pwd)) stop("pwd (password) entered is not a string")
+  if (!is.character(assetid)) stop("assetid entered is not a string")
+  if (is.null(url) | url == "") stop("URL empty")
+  if (is.null(uname) | uname == "") stop("uname (username) empty")
+  if (is.null(pwd) | pwd == "") stop("pwd (password) empty")
+  if (is.null(assetid) | assetid == "") stop("assetid empty")
+  if (!format %in% c("json", "xml")) stop("format must be either 'json' or 'xml'")
+
+  base_url <- paste0("https://", url, "/api/v2/assets/", assetid, "/data.", format)
+
+  message("Starting data retrieval from ", base_url)
+
+  get_page <- function(url, limit = 30000, start = 0) {
+    full_url <- paste0(url, "?limit=", limit, "&start=", start)
+
+    message("Retrieving page starting at record ", start)
+
+    respon.kpi <- tryCatch(
+      expr = {
+        httr2::request(full_url) |>
+          httr2::req_auth_basic(uname, pwd) |>
+          httr2::req_perform()
+      },
+      error = function(x) {
+        message("Error on page starting at record ", start, ". Please try again or check the input parameters.")
+        return(NULL)
+      }
+    )
+
+    if (!is.null(respon.kpi)) {
+      content_type <- httr2::resp_content_type(respon.kpi)
+
+      if (grepl("json", content_type)) {
+        message("Successfully retrieved JSON data starting at record ", start)
+        return(httr2::resp_body_json(respon.kpi, encoding = encoding))
+      } else if (grepl("xml", content_type)) {
+        message("Successfully retrieved XML data starting at record ", start)
+        return(httr2::resp_body_string(respon.kpi, encoding = encoding))
+      } else if (grepl("html", content_type)) {
+        warning("Unexpected HTML response for start ", start, ". Unable to parse.")
+        return(NULL)
+      } else {
+        warning("Unexpected content type: ", content_type, " for start ", start, ". Unable to parse.")
+        return(NULL)
+      }
+    } else {
+      return(NULL)
+    }
+  }
+
+  all_results <- list()
+  start <- 0
+  limit <- 30000
+  get_next <- TRUE
+
+  while (get_next) {
+    page_results <- get_page(base_url, limit, start)
+
+    if (is.null(page_results)) {
+      message("Error occurred. Stopping data retrieval.")
+      break
+    }
+
+    new_results <- page_results$results
+    all_results <- c(all_results, new_results)
+
+    message("Total records retrieved so far: ", length(all_results))
+
+    if (length(new_results) < limit) {
+      message("Retrieved all available records.")
+      get_next <- FALSE
+    } else {
+      start <- start + limit
+    }
+  }
+
+  message("Data retrieval complete. Total records retrieved: ", length(all_results))
+
+  # Check for unique submission IDs
+  submission_ids <- sapply(all_results, function(x) x$`_id`)
+  if (length(unique(submission_ids)) != length(all_results)) {
+    warning("Number of unique submission IDs does not match the number of records. There may be duplicates.")
+  }
+
   data_filename <- paste(prefix, "raw", sep = "_")
+  csv_filename <- add_version(data_filename, "csv")
 
-  if (isTRUE(append_version)) {
-    metadata_filename <- add_version(metadata_filename, "json")
-    csv_filename <- add_version(data_filename, "csv")
-    json_filename <- add_version(data_filename, "json")
-  } else {
-    metadata_filename <- paste0(metadata_filename, ".json")
-    csv_filename <- paste0(data_filename, ".csv")
-    json_filename <- paste0(data_filename, ".json")
-  }
-
-  filenames <- character()
-
-  if (isTRUE(metadata)) {
-    logger::log_info("Downloading survey metadata as {metadata_filename}...")
-    retrieve_survey_metadata(id, token = token, api = api) %>%
-      jsonlite::write_json(metadata_filename)
-    logger::log_success("Metadata download succeeded")
-    filenames <- c(filenames, metadata_filename)
-  }
-
-  logger::log_info("Downloading survey json data as {json_filename}...")
-  retrieve_survey_data(json_filename, id, token, api)
-  logger::log_success("Survey json data download succeeded")
-
-  if ("csv" %in% format) {
-    logger::log_info("Converting json data to CSV as {csv_filename}...")
-    survey_json_to_csv(json_filename, csv_filename)
-    filenames <- c(filenames, csv_filename)
-  }
-
-  if ("json" %in% format) {
-    filenames <- c(filenames, json_filename)
-  } else {
-    logger::log_info("Removing temporary json {json_filename}...")
-    file.remove(json_filename)
-  }
-
-  filenames
+  logger::log_info("Converting json data to CSV as {csv_filename}...")
+  survey_df <- purrr::map_dfr(all_results, flatten_row)
+  readr::write_csv(survey_df, csv_filename)
+  csv_filename
 }
 
-# Convert a Kobo survey in json format to a csv format that respects the
-# conventions from the data that could have been downloaded using the csv
-# request in the API v1
-survey_json_to_csv <- function(json_path, csv_path) {
-  survey <- jsonlite::read_json(json_path)
-  survey_df <- purrr::map_dfr(survey, flatten_row)
-  readr::write_csv(survey_df, csv_path)
-  csv_path
-}
+
 
 flatten_row <- function(x) {
   x %>%